pyproject.toml
[project]
name = "ai-workshop"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.124.0",
    "portkey-ai>=2.1.0",
    "python-dotenv>=1.2.1",
    "retell-sdk>=5.1.0",
    "uvicorn>=0.38.0",
]
